requirements.txt
gradio==4.31.3
langchain==0.1.20
langchain-openai==0.1.7
pgvector==0.2.5 
psycopg2==2.9.9
python-dotenv==1.0.1
uvicorn==0.29.0
langchain-text-splitters==0.0.2